decomp patch: (basename ../../patches/decomp/045-shprogrscan-ulong-to-u32.patch .patch)

+++ src/Game/SH/SHProgressiveScan.cpp
@@ -55,7 +55,7 @@ ProgressiveScanScene::~ProgressiveScanScene()
     delete mConfirmationImage;
 }
 
-extern unsigned long nlStringLowerHash(const char*);
+extern u32 nlStringLowerHash(const char*);
 extern int nlSNPrintf(char*, unsigned long, const char*, ...);
 
 /**
